modify metrics test
